% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{arrows is overlay in plot}
\usage{
benchmark(cpu = T, cal = F, na_rm = F, amd = F, intel = F,
  money_keep = F, order, header = 0, up = 0)
}
\arguments{
\item{money_keep}{if 'money_keep =T' then '1 banchmark point of price' is cbind}

\item{order}{sort number if money_keep=T then order=1~4 is range. else if money_keep=F then order=1~3.}

\item{up}{is 'up' grater than banchmark score output.}

\item{cpu=}{T or F. T is output of cpu , F is output of gpu}

\item{cal=}{delete of comma}

\item{amd=}{if amd=T only output of AMD cpu.}

\item{intal=}{if intel=T only output of intel cpu.}
}
\description{
arrows is overlay in plot
}
\examples{
benchmark(cpu=F,amd=T,header=0,na_rm=F,money_keep = T,order=4,up=500)
}
